code reformate, small package and rollup changes
