+++ package.json
@@ -5,9 +5,9 @@
     "main": "main.js",
     "type":"module",
     "scripts": {
-        "dev": "set NODE_ENV=development && npx rollup --configPlugin @rollup/plugin-commonjs,@rollup/plugin-node-resolve,@rollup/plugin-typescript -c  rollup.config.mjs -w && sass --watch ./src/Styles/global.scss ./dist/styles.css --no-source-map",
-        "build": "set NODE_ENV=production & npx rollup --configPlugin @rollup/plugin-commonjs,@rollup/plugin-node-resolve,@rollup/plugin-typescript -c  rollup.config.mjs & sass ./src/Styles/global.scss ./dist/styles.css --no-source-map",
-        "test": "npx rollup --config rollup.config.mjs & sass ./src/Styles/global.scss ./styles.css --no-source-map"
+        "dev": "npx rollup --configPlugin @rollup/plugin-commonjs,@rollup/plugin-node-resolve,@rollup/plugin-typescript -c  rollup.config.mjs -w & sass --watch ./src/Styles/global.scss ./dist/styles.css --no-source-map",
+        "build": "npx rollup --configPlugin @rollup/plugin-commonjs,@rollup/plugin-node-resolve,@rollup/plugin-typescript -c  rollup.config.mjs && sass ./src/Styles/global.scss ./dist/styles.css --no-source-map",
+        "test": "npx rollup --config rollup.config.mjs & sass ./src/Styles/global.scss ./dist/styles.css --no-source-map"
     },
     "keywords": [],
     "author": "ReconVirus",
