+++ rollup.config.mjs
@@ -3,12 +3,10 @@ import {nodeResolve} from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import styles from "rollup-plugin-styles";
 
-const isDev = process.env.NODE_ENV === 'development';
-
 export default {
     input: './src/Main.ts',
     output: {
-        dir: isDev ? '.' : './dist',
+        dir: './dist',
         sourcemap: 'inline',
         format: 'cjs',
         exports: 'default'
