+++ src/Block.ts
@@ -1,8 +1,9 @@
 import type {FrontmatterKeys, NoteData, TimelineArgs, HistoriumSettings} from './Types';
 import {BST} from './BTS';
+import {getFrontmatterData} from './Frontmatter';
 import {createDate, FilterMDFiles, getImgUrl, parseTag} from './Utils';
 import {VerticalTimeline} from './VerticalTimeline'
-import {FrontMatterCache, MarkdownView, MetadataCache, Notice, TFile, Vault,} from 'obsidian'
+import {MarkdownView, MetadataCache, TFile, Vault,} from 'obsidian'
 import {DataSet} from "vis-data";
 import {Timeline} from "vis-timeline/esnext";
 import "vis-timeline/styles/vis-timeline-graph2d.css";
@@ -270,29 +271,4 @@ async function getTimelineData(appVault: Vault, fileCache: MetadataCache, fileLi
         }
     }
     return [timelineNotes, Array.from(timelineDates.inOrder())];
-}
-function getFrontmatterData(frontmatter: FrontMatterCache | null, frontmatterKeys: FrontmatterKeys, file: TFile): [string, string, string, string, string, string, string, string, string, string | null] {
-	const startDate = findMatchingFrontmatterKey(frontmatter, frontmatterKeys.startDateKey);
-	if (!startDate) {
-		new Notice(`No date found for ${file.name}`);
-		return ['', '', '', '', '', '', '', '', '', ''];
-	}
-	const noteTitle = findMatchingFrontmatterKey(frontmatter, frontmatterKeys.titleKey) ?? file.name.replace(".md", "");
-	const noteDescription = frontmatter?.description;
-	const noteImage = frontmatter?.image;
-    const noteIndicator = findMatchingFrontmatterKey(frontmatter, frontmatterKeys.indicatorKey);
-    const type = frontmatter["type"] ?? 'box';
-	const noteClass = frontmatter["color"] ?? '';
-	const notePath = '/' + file.path;
-	const endDate = findMatchingFrontmatterKey(frontmatter, frontmatterKeys.endDateKey) ?? null;
-    const noteGroup = frontmatter?.group;
-	return [startDate, noteTitle, noteDescription, noteImage, noteIndicator, type, noteClass, notePath, endDate, noteGroup];
-}
-function findMatchingFrontmatterKey(frontmatter: FrontMatterCache | null, keys: string[]) {
-	for (const key of keys) {
-		if (frontmatter && frontmatter[key]) {
-			return frontmatter[key];
-		}
-	}
-	return null;
 }
